Разбиваем большой файл с unit-тестами
